Moved plugin from conftest to its own module
